Refactor: Moved components around for better organisation
