+++ frontend/src/components/private/Composer.jsx
@@ -4,8 +4,8 @@ import {
   updatePost,
   updateComment,
   createPostComment,
-} from "../api/postService.js";
-import useAuth from "../hooks/useAuth.js";
+} from "../../api/postService.js";
+import useAuth from "../../hooks/useAuth.js";
 import DOMPurify from "dompurify";
 import {
   MDXEditor,
