+++ frontend/src/components/private/Post.jsx
@@ -1,10 +1,10 @@
 import { useEffect, useState } from "react";
-import getTimeAgo from "../util/getTimeAgo";
+import getTimeAgo from "../../util/getTimeAgo";
 import { FaCommentDots, FaHeart, FaShare, FaLink } from "react-icons/fa6";
-import { likePost, getPostComments, deletePostById } from "../api/postService";
-import DropdownMenu from "./DropdownMenu";
+import { likePost, getPostComments, deletePostById } from "../../api/postService";
+import DropdownMenu from "../shared/DropdownMenu";
 import Composer from "./Composer";
-import { useCacheUpdater, useCachedUser } from "../hooks/useUserCache";
+import { useCacheUpdater, useCachedUser } from "../../hooks/useUserCache";
 import Comment from "./Comment";
 import Markdown from "react-markdown";
 import { Popover } from "flowbite-react";
