+++ frontend/src/layouts/PublicLayout.jsx
@@ -1,5 +1,5 @@
 import { Outlet } from "react-router-dom";
-import Navbar from "../components/Navbar";
+import Navbar from "../components/public/Navbar";
 
 export default function PublicLayout() {
   return (
